Add unittest for Python backend environment (#2174)
* Add unittest for Python backend environment variable propagation

* Review edits

+++ qa/L0_backend_python/test.sh
@@ -87,7 +87,7 @@ fi
 RET=0
 
 set +e
-python $CLIENT_PY >>$CLIENT_LOG 2>&1
+python3 $CLIENT_PY >>$CLIENT_LOG 2>&1
 if [ $? -ne 0 ]; then
     RET=1
 else
@@ -180,6 +180,7 @@ if [ $? -ne 0 ]; then
 fi
 set -e
 
+# Test KIND_GPU
 rm -rf models/
 mkdir -p models/add_sub_gpu/1/
 cp ../python_models/add_sub/model.py ./models/add_sub_gpu/1/
@@ -192,11 +193,28 @@ if [ "$SERVER_PID" == "0" ]; then
     exit 1
 fi
 
-
 if [ $? -ne 0 ]; then
     cat $CLIENT_LOG
     echo -e "\n***\n*** KIND_GPU model test failed \n***"
     RET=1
+fi
+
+kill $SERVER_PID
+wait $SERVER_PID
+
+# Test environment variable propagation
+rm -rf models/
+mkdir -p models/model_env/1/
+cp ../python_models/model_env/model.py ./models/model_env/1/
+cp ../python_models/model_env/config.pbtxt ./models/model_env/
+
+export MY_ENV="MY_ENV"
+run_server
+if [ "$SERVER_PID" == "0" ]; then
+    echo -e "\n***\n*** Failed to start $SERVER\n***"
+    echo -e "\n***\n*** Environment variable test failed \n***"
+    cat $SERVER_LOG
+    exit 1
 fi
 
 kill $SERVER_PID
